+++ src/main/java/org/example/userauthenticationservice/UserAuthenticationServiceApplication.java
@@ -1,0 +1,13 @@
+package org.example.userauthenticationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserAuthenticationServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserAuthenticationServiceApplication.class, args);
+    }
+
+}
