+++ src/main/java/org/example/userauthenticationservice/exceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package org.example.userauthenticationservice_sept2024.exceptions;
+package org.example.userauthenticationservice.exceptions;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
