+++ src/main/java/org/example/userauthenticationservice/models/RoleName.java
@@ -1,0 +1,5 @@
+package org.example.userauthenticationservice.models;
+
+public enum RoleName {
+    ADMIN
+}
