+++ src/main/java/org/example/userauthenticationservice/models/UserAuthenticationStatus.java
@@ -1,4 +1,4 @@
-package org.example.userauthenticationservice_sept2024.models;
+package org.example.userauthenticationservice.models;
 
 public enum UserAuthenticationStatus {
     SUCCESS,
