--- src/main/java/org/example/userauthenticationservice_sept2024/UserAuthenticationServiceSept2024Application.java
@@ -1,13 +1,0 @@
-package org.example.userauthenticationservice_sept2024;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserAuthenticationServiceSept2024Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserAuthenticationServiceSept2024Application.class, args);
-    }
-
-}
