--- src/main/java/org/example/userauthenticationservice_sept2024/models/RoleName.java
@@ -1,5 +1,0 @@
-package org.example.userauthenticationservice_sept2024.models;
-
-public enum RoleName {
-    ADMIN
-}
