+++ src/test/java/org/example/userauthenticationservice/UserAuthenticationServiceSept2024ApplicationTests.java
@@ -1,4 +1,4 @@
-package org.example.userauthenticationservice_sept2024;
+package org.example.userauthenticationservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
